cheri: Fix suspected upstream bug in sound/soc/codecs/ab8500-codec.c
The CHERI clang-tidy checker complains about incompatible pointer
target types in the target of a cast in ab8500_codec_probe().

Apparently, the code is in fact wrong: The "private_value" field
is a pointer to a "struct filter_control" encoded as a uintptr_t.
This matches the code in the initialization. However, when using
the "private_value" its address and not its contents are cast to
"struct filter_control".

Given that the code is there since 2012 it must be either mostly
unused or I am missing something.

Thus keep this as a separate commit for the time being.

+++ sound/drivers/mpu401/mpu401_uart.c
@@ -233,7 +233,7 @@ static int snd_mpu401_uart_cmd(struct snd_mpu401 * mpu, unsigned char cmd,
 	if (!ok) {
 		dev_err(mpu->rmidi->dev,
 			"cmd: 0x%x failed at 0x%lx (status = 0x%x, data = 0x%x)\n",
-			cmd, mpu->port,
+			cmd, (unsigned long)mpu->port,
 			mpu->read(mpu, MPU401C(mpu)),
 			mpu->read(mpu, MPU401D(mpu)));
 		return 1;
@@ -537,7 +537,7 @@ int snd_mpu401_uart_new(struct snd_card *card, int device,
 		if (!mpu->res) {
 			dev_err(rmidi->dev,
 				"mpu401_uart: unable to grab port 0x%lx size %d\n",
-				port, res_size);
+				(unsigned long)port, res_size);
 			err = -EBUSY;
 			goto free_device;
 		}
